app/requirements.txt
flask===2.1.2
flask-cors
psycopg2
Flask-SQLAlchemy
Flask-Migrate
Flask-Marshmallow
marshmallow-sqlalchemy